requirements.txt
Flask #python web frakework provides local server to host webpages
requests #module
python-dotenv #lib